Revert "Move utility method from evaluation utilities class to core utilities class."
This reverts commit 891b13c37e20a8b4e7d7ef308da7d9fafcb6a41f.
